+++ cli/src/common/boto3/ec2.py
@@ -8,7 +8,7 @@
 # or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
-from common.aws.aws_resources import InstanceInfo
+from common.aws.aws_resources import ImageInfo, InstanceInfo
 from common.boto3.common import AWSClientError, AWSExceptionHandler, Boto3Client
 from pcluster import utils
 from pcluster.utils import Cache, InstanceTypeInfo
@@ -48,11 +48,19 @@ class Ec2Client(Boto3Client):
 
     @AWSExceptionHandler.handle_client_exception
     def describe_image(self, ami_id):
-        """Return a dict of ami info."""
+        """Describe image by image id, return a dict of ami info."""
         result = self._client.describe_images(ImageIds=[ami_id])
         if result.get("Images"):
             return result.get("Images")[0]
         raise AWSClientError(function_name="describe_image", message=f"Image {ami_id} not found")
+
+    @AWSExceptionHandler.handle_client_exception
+    def describe_images(self, filters=None, owners=None):
+        """Describe images by filters and owners, return a list of ImageInfo."""
+        return [
+            ImageInfo(image_info)
+            for image_info in self._client.describe_images(Filters=filters, Owners=owners).get("Images")
+        ]
 
     @AWSExceptionHandler.handle_client_exception
     def describe_key_pair(self, key_name):