+++ cli/src/pcluster/cli.py
@@ -82,6 +82,10 @@ def build_image(args):
     pcluster.build_image(args)
 
 
+def list_images(args):
+    pcluster.list_images(args)
+
+
 def config_logger():
     logger = logging.getLogger("pcluster")
     file_only_logger = logging.getLogger("cli_log_file")
@@ -295,7 +299,7 @@ Returns an ssh command with the cluster username and IP address pre-populated::
     pssh.add_argument("-d", "--dryrun", action="store_true", default=False, help="Prints command and exits.")
     pssh.set_defaults(func=ssh)
 
-    # createami command subparser
+    # build image command subparser
     pami = subparsers.add_parser("build-image", help="Creates a custom AMI to use with AWS ParallelCluster.")
     pami.add_argument(
         "-i",
@@ -307,6 +311,18 @@ Returns an ssh command with the cluster username and IP address pre-populated::
     _addarg_config(pami)
     _addarg_region(pami)
     pami.set_defaults(func=build_image)
+
+    # list images command subparser
+    plistami = subparsers.add_parser(
+        "list-images",
+        help="Displays a list of images built by AWS ParallelCluster in a given AWS region associated with "
+        "status and version.",
+        epilog="This command lists the name, status and version of images built by AWS ParallelCluster in a given "
+        "AWS region.",
+    )
+    plistami.add_argument("--color", action="store_true", default=False, help="Display the cluster status in color.")
+    _addarg_region(plistami)
+    plistami.set_defaults(func=list_images)
 
     # configure command subparser
     pconfigure = subparsers.add_parser("configure", help="Start the AWS ParallelCluster configuration.")