+++ cli/src/pcluster/models/imagebuilder.py
@@ -78,14 +78,14 @@ class ImageBuilderStack(StackInfo):
     @property
     def version(self):
         """Return the version of ParallelCluster used to create the stack."""
-        return self._get_tag("pcluster_version")
+        return self._get_tag("pcluster_build_image")
 
     @property
     def image_id(self):
         """Return output image id."""
         if self._image_resource:
             try:
-                image_build_version_arn = self._image_resource.get("StackResourceDetail").get("PhysicalResourceId")
+                image_build_version_arn = self._image_resource["StackResourceDetail"]["PhysicalResourceId"]
                 return AWSApi.instance().imagebuilder.get_image_id(image_build_version_arn)
             except (AWSClientError, KeyError):
                 return None
@@ -95,7 +95,8 @@ class ImageBuilderStack(StackInfo):
     def get_source_config(self):
         """Get source config from metadata."""
         try:
-            return AWSApi.instance().cfn.get_template(self.name).get("Metadata").get("Config")
+            template_body = AWSApi.instance().cfn.get_stack_template(self.name)
+            return template_body["Metadata"]["Config"] if template_body else None
         except (AWSClientError, KeyError):
             return None
 
@@ -103,7 +104,7 @@ class ImageBuilderStack(StackInfo):
 class ImageBuilder:
     """Represent a building image, composed by an ImageBuilder config and an ImageBuilderStack."""
 
-    def __init__(self, image_name: str, config: dict = None, stack: ImageBuilderStack = None):
+    def __init__(self, image_name: str = None, config: dict = None, stack: ImageBuilderStack = None):
         self.image_name = image_name
         self.__source_config = config
         self.__stack = stack